# docs/swagger.yaml
basePath: /api
definitions:
  models.Pomodoro:
    properties:
      created_at:
        type: string
      date:
        type: string
      duration:
        type: integer
      id:
        type: integer
      user:
        $ref: '#/definitions/models.User'
      user_id:
        type: integer
    type: object
  models.Task:
    properties:
      created_at:
        type: string
      deadline:
        type: string
      description:
        type: string
      id:
        type: integer
      is_complete:
        type: boolean
      priority:
        description: 0=rendah, 1=sedang, 2=tinggi,3=paling tinggi
        type: integer
      title:
        type: string
      todo:
        $ref: '#/definitions/models.Todo'
      todo_id:
        type: integer
      updated_at:
        type: string
    type: object
  models.Todo:
    properties:
      created_at:
        type: string
      id:
        type: integer
      is_complete:
        type: boolean
      priority:
        description: 0=rendah, 1=sedang, 2=tinggi,3=paling tinggi
        type: integer
      progress:
        type: integer
      tasks:
        items:
          $ref: '#/definitions/models.Task'
        type: array
      title:
        type: string
      updated_at:
        type: string
      user:
        $ref: '#/definitions/models.User'
      user_id:
        type: integer
    type: object
  models.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      google_id:
        type: string
      id:
        type: integer
      name:
        type: string
      phone:
        type: string
      picture:
        type: string
      pomodoros:
        items:
          $ref: '#/definitions/models.Pomodoro'
        type: array
      provider:
        type: string
      todos:
        items:
          $ref: '#/definitions/models.Todo'
        type: array
      updated_at:
        type: string
      username:
        type: string
    type: object
host: api.tascaid.com
info:
  contact: {}
  description: API untuk aplikasi Tasca A4 AGILE menggunakan gin
  title: Tag Tasca API
  version: "1.0"
paths:
...
  /api/todos:
    get:
      description: Get all todos for the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: 'data: []models.Todo'
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 'error: Unauthorized'
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 'error: Gagal fetch data'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get all todos
      tags:
      - todos
    post:
      consumes:
      - application/json
      description: Create a new todo
      parameters:
      - description: Todo data
        in: body
        name: todo
        required: true
        schema:
          $ref: '#/definitions/models.Todo'
      produces:
      - application/json
      responses:
        "201":
          description: 'message: Todo berhasil dibuat'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: 'error: Input tidak valid'
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: 'error: Unauthorized'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Create todo
      tags:
      - todos
  /api/todos/{id}:
    delete:
      description: Delete a todo
      parameters:
      - description: Todo ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 'message: Todo berhasil dihapus'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: 'error: ID tidak valid'
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: 'error: Unauthorized'
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: 'error: Todo tidak ditemukan atau bukan milik Anda'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Delete todo
      tags:
      - todos
    get:
      description: Get a specific todo by ID
      parameters:
      - description: Todo ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 'data: models.Todo'
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 'error: ID tidak valid'
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: 'error: Unauthorized'
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: 'error: Todo tidak ditemukan'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get todo by ID
      tags:
      - todos
    patch:
      consumes:
      - application/json
      description: Update an existing todo
      parameters:
      - description: Todo ID
        in: path
        name: id
        required: true
        type: integer
      - description: Todo update data
        in: body
        name: todo
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 'message: Todo updated successfully'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: 'error: Invalid todo ID'
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: 'error: Unauthorized'
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: 'error: Tidak memiliki akses ke todo ini'
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 'error: Update failed'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Update todo
      tags:
      - todos
...
swagger: "2.0"